refactor: RoomUser에 unreadCount, lastMessageContent, lastMessageTime 컬럼 추가

+++ src/main/java/com/yuguanzhang/lumi/chat/dto/ChatRoomsResponseDto.java
@@ -16,5 +16,5 @@ public class ChatRoomsResponseDto {
     private final String opponentName;
     private final LocalDateTime lastMessageTime;
     private final String lastMessage;
-    private final Long unreadMessageCount;
+    private final int unreadMessageCount;
 }
